We now onyl attempt to delte and restore the directory if it existed. This is perfroemd through the addition of the "directory_deleted" flag.
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@36778 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/build2/commands/gb_command_delete_directory.e
@@ -57,7 +57,13 @@ feature -- Basic Operation
 			create temp_file_name.make_from_string (generated_path.string)
 			temp_file_name.extend (directory_name)	
 			create directory.make (temp_file_name)
-			delete_directory (directory)
+			if directory.exists then
+					-- Only remove the directory if it is present on the disk.
+					-- If a project has not been generated, then there may be no directory yet.
+				delete_directory (directory)
+					-- Record the deletion
+				directory_deleted := True
+			end
 			if not history.command_list.has (Current) then
 				history.add_command (Current)
 			end
@@ -72,14 +78,18 @@ feature -- Basic Operation
 			temp_file_name: FILE_NAME
 			directory: DIRECTORY
 		do
-			create temp_file_name.make_from_string (generated_path.string)
-			temp_file_name.extend (directory_name)	
-			create directory.make (temp_file_name)
-			create_directory (directory)
+			if directory_deleted then
+					-- Only restore the directory to the disk if it was actually
+					-- removed previously.
+				create temp_file_name.make_from_string (generated_path.string)
+				temp_file_name.extend (directory_name)	
+				create directory.make (temp_file_name)
+				create_directory (directory)
+			end
 			window_selector.add_named_directory (directory_name)
 			command_handler.update
 		end
-		
+	
 	textual_representation: STRING is
 			-- Text representation of command exectuted.
 		do
@@ -99,5 +109,8 @@ feature {NONE} -- Implementation
 	directory_name: STRING
 		-- Name of directory that was deleted. Only the actual name,
 		-- and not the full name, as all directories are relative to the project.
+		
+	directory_deleted: BOOLEAN
+		-- Was directory `directory_name' actually deleted from the disk?
 
 end -- class GB_COMMAND_DELETE_DIRECTORY
